+++ spec/connection_throttle_spec.rb
@@ -4,7 +4,9 @@ module Ftpd
 
   describe ConnectionThrottle do
 
+    let(:socket) {mock TCPSocket}
     let(:connections) {0}
+    let(:connections_for_socket) {0}
     let(:connection_tracker) do
       mock ConnectionTracker, :connections => connections
     end
@@ -12,31 +14,68 @@ module Ftpd
       ConnectionThrottle.new(connection_tracker)
     end
 
+    before(:each) do
+      connection_tracker.stub(:connections => connections)
+      connection_tracker.stub(:connections_for).with(socket).and_return(connections_for_socket)
+    end
+
     it 'should have defaults' do
       connection_throttle.max_connections.should == 100
+      connection_throttle.max_connections_per_ip.should == 5
     end
 
     describe '#allow?' do
-      
-      let(:max_connections) {50}
-      let(:socket) {mock TCPSocket}
-      before(:each) do
-        connection_throttle.max_connections = max_connections
+
+      context '(total connections)' do
+        
+        let(:max_connections) {50}
+
+        before(:each) do
+          connection_throttle.max_connections = max_connections
+          connection_throttle.max_connections_per_ip = 2 * max_connections
+        end
+
+        context 'almost at maximum connections' do
+          let(:connections) {max_connections - 1}
+          specify {connection_throttle.allow?(socket).should be_true}
+        end
+
+        context 'at maximum connections' do
+          let(:connections) {max_connections}
+          specify {connection_throttle.allow?(socket).should be_false}
+        end
+
+        context 'above maximum connections' do
+          let(:connections) {max_connections + 1}
+          specify {connection_throttle.allow?(socket).should be_false}
+        end
+
       end
 
-      context 'almost at maximum connections' do
-        let(:connections) {max_connections - 1}
-        specify {connection_throttle.allow?(socket).should be_true}
-      end
+      context '(per ip)' do
+        
+        let(:max_connections_per_ip) {5}
 
-      context 'at maximum connections' do
-        let(:connections) {max_connections}
-        specify {connection_throttle.allow?(socket).should be_false}
-      end
+        before(:each) do
+          connection_throttle.max_connections = 2 * max_connections_per_ip
+          connection_throttle.max_connections_per_ip = max_connections_per_ip
+        end
 
-      context 'above maximum connections' do
-        let(:connections) {max_connections + 1}
-        specify {connection_throttle.allow?(socket).should be_false}
+        context 'almost at maximum connections for ip' do
+          let(:connections_for_socket) {max_connections_per_ip - 1}
+          specify {connection_throttle.allow?(socket).should be_true}
+        end
+
+        context 'at maximum connections for ip' do
+          let(:connections_for_socket) {max_connections_per_ip}
+          specify {connection_throttle.allow?(socket).should be_false}
+        end
+
+        context 'above maximum connections for ip' do
+          let(:connections_for_socket) {max_connections_per_ip + 1}
+          specify {connection_throttle.allow?(socket).should be_false}
+        end
+
       end
 
     end